+++ CHANGELOG.md
@@ -14,6 +14,7 @@ details.
 ## [Unreleased]
 ### Added
 - Extended configuration file for [Preact](https://preactjs.com/) based projects.
+- All published rules as of ESLint 3.15.0
 
 ## [1.4.0] - 2017-01-21
 ### Added